First pass at adding graphics to the game.  Game#show displays one board an aircraft carrier, white pegs in all the holes with their positions adjusted to account for the skew of the board image and red pegs on the four corners and along all of the ship.  The helpers/games_helper.rb has a method to calculate the left and top absolute positioning of a peg based on the target coordinate.
Next steps are to pass in the player id (or later use devise for loging
in, user authentication, and thus know current_user's id).  for a player
have the controller pass in the coordinates of the player's ships and
the set of shots fired at the player so far and to display the ships and
white or red pegs at the target locations.
Then to display the targetting board and the shots this player fired,
and I think the identity of known sunk ships.